+++ citydb-cli/src/main/java/org/citydb/cli/index/options/OutputOptions.java
@@ -19,26 +19,24 @@
  * limitations under the License.
  */
 
-package org.citydb.cli.common;
+package org.citydb.cli.index.options;
 
 import picocli.CommandLine;
 
 import java.nio.file.Path;
 
-public class OutputFileOptions implements Option {
-    @CommandLine.Option(names = {"-o", "--output"}, required = true,
-            description = "Name of the output file.")
+public class OutputOptions extends org.citydb.cli.common.OutputOptions {
+    @CommandLine.Option(names = {"-o", "--output"}, paramLabel = "<file|->",
+            description = "Write index status to JSON. Use '-' for stdout.")
     private Path file;
 
-    @CommandLine.Option(names = "--output-encoding",
-            description = "Encoding to use for the output file.")
-    private String encoding;
-
+    @Override
     public Path getFile() {
         return file;
     }
 
-    public String getEncoding() {
-        return encoding;
+    @Override
+    protected void setFile(Path file) {
+        this.file = file;
     }
 }